Fix SMS timezone reading/writing to/from pdu
The timezone (de)serialization was incorrect. Before e159509 all
timezones because of a bug were treated as positive when read.

In e159509 the issue was partially fixed. But the sign was reversed: 1
in 3rd bit meant that timezone is positive and 0 that it's negative when
GSM spec says it should be vice-versa.

See: https://en.wikipedia.org/wiki/GSM_03.40#Time_Format